feat(release): ground-up rewrite and some changes to the public API
This commit represents a full rewrite of the plugin. With it comes better configurability, performance, and better interoperability between Typescript, Babel, Browserslists, and Rollup.
Some highlights include:
- Config files will now be discovered automatically
- Declaration files will now work seamlessly with code splitting
- Diagnostics will be emitted via Rollup's PluginContext API and presented using Rollup's RollupWarning and RollupError interfaces
- CustomTransformers are now supported
- Intelligent defaults for Babel will be used with deduping of plugins and presets.
- Intelligent handling of minification presets and plugins for Babel such that they will always run per-chunk rather than per-file like all other provided Babel plugins and presets.

Closes #2

BREAKING CHANGE: The 'babel' config option has been removed in favor of 'babelConfig' which may point to a file on disk (such as a '.babelrc' or a 'babel.config.js') or be a dictionary of Babel options.
BREAKING CHANGE: The 'parseExternalModules' config option has been removed.